pyproject.toml
[tool.poetry]
name = "CineReserve"
version = "0.1.0"
description = "Cinema reservation API with JWT, Redis locks, and Celery"
authors = ["franklinclf <franklin.oliveira.049@ufrn.edu.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "5.1.5"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.3.1"
drf-spectacular = "^0.29.0"
drf-spectacular-sidecar = "^2026.3.1"
redis = "^7.3.0"
hiredis = "^3.3.1"
django-redis = "^6.0.0"
psycopg = "^3.3.3"
psycopg-binary = "^3.3.3"
flower = "^2.0.1"
celery = "^5.4.0"
django-cors-headers = "^4.9.0"
django-environ = "^0.13.0"
gunicorn = "^23.0.0"
whitenoise = "^6.8.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.9.0"
pytest-cov = "^6.0.0"
factory-boy = "^3.3.1"
model-bakery = "^1.20.0"
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
safety = "^3.2.11"
bandit = "^1.7.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
